Base Graph class with to_dot
